requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
certifi==2022.12.7
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.2
django-cors-headers==3.14.0
django-debug-toolbar==4.0.0
django-rest-multiple-models==2.1.3
django-swagger-ui==0.1.11
djangorestframework==3.14.0
drf-access-policy==1.5.0
drf-yasg==1.21.5
future==0.18.3
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
packaging==23.1
psycopg2-binary==2.9.6
pyparsing==3.0.9
pytz==2023.3
PyYAML==6.0
requests==2.28.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
serializers==0.2.4
sqlparse==0.4.4
uritemplate==4.1.1
urllib3==1.26.15
asgiref==3.6.0
backports.zoneinfo==0.2.1
certifi==2022.12.7
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.2
django-cors-headers==3.14.0
django-debug-toolbar==4.0.0
django-rest-multiple-models==2.1.3
django-swagger-ui==0.1.11
djangorestframework==3.14.0
drf-access-policy==1.5.0
drf-yasg==1.21.5
future==0.18.3
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
packaging==23.1
psycopg2-binary==2.9.6
pyparsing==3.0.9
pytz==2023.3
PyYAML==6.0
requests==2.28.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
serializers==0.2.4
sqlparse==0.4.4
uritemplate==4.1.1
urllib3==1.26.15
asgiref==3.6.0
backports.zoneinfo==0.2.1
certifi==2022.12.7
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.2
django-cors-headers==3.14.0
django-debug-toolbar==4.0.0
django-rest-multiple-models==2.1.3
django-swagger-ui==0.1.11
djangorestframework==3.14.0
drf-access-policy==1.5.0
drf-yasg==1.21.5
future==0.18.3
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
packaging==23.1
psycopg2-binary==2.9.6
pyparsing==3.0.9
pytz==2023.3
PyYAML==6.0
requests==2.28.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
serializers==0.2.4
sqlparse==0.4.4
uritemplate==4.1.1
urllib3==1.26.15
